@@ -4,45 +4,43 @@ namespace App\Controller;
 
 use App\Entity\RegExpedientes;
 use App\Form\RegExpedienteSearchFormType;
+use App\Repository\PasosExpedientesRepository;
 use App\Repository\RegExpedientesRepository;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Translation\TranslatableMessage;
+use \Symfony\Component\Security\Http\Attribute\IsGranted;
 
+#[IsGranted('ROLE_ESPEDIENTEAK')]
 class RegExpedientesController extends BaseController
 {
-    private RegExpedientesRepository $repo;
-    private string $urlDocumentos;
-    private int $maxExpedientes;
-
-    public function __construct(RegExpedientesRepository $repo, $urlDocumentos = '/documentos', $maxExpedientes = 50, $parameters)
+    public function __construct(
+        private readonly PasosExpedientesRepository $pasosRepo, 
+        private readonly RegExpedientesRepository $repo,
+        $parameters, 
+        private readonly string $urlDocumentos = '/documentos', 
+        private readonly int $maxExpedientes = 50)
     {
-        $this->repo = $repo;
-        $this->urlDocumentos = $urlDocumentos;
-        $this->maxExpedientes = $maxExpedientes;
         parent::__construct($parameters);
     }
 
-    /**
-     * @Route("/{_locale}/regexpedientes/{regexpediente}/detail", name="regexpendientes_detail")
-     */
+    #[Route(path: '/{_locale}/regexpedientes/{regexpediente}/detail', name: 'regexpendientes_detail')]
     public function show(Request $request, RegExpedientes $regexpediente): Response
     {
         $this->loadQueryParameters($request);
         $criteria = $request->query->all();
         $criteria = $this->removePaginationParameters($criteria);
-        $pasosexpedientes = $regexpediente->getPasosExpedientes();
+        $pasosexpedientes = $this->pasosRepo->getPasosExpendienteOrdered($regexpediente);
         $template = !$this->getAjax() ? 'regexpedientes/show.html.twig' : 'regexpedientes/_pasosList.html.twig';
         return $this->renderForm($template, [
             'pasosexpedientes' => $pasosexpedientes,
             'urlDocumentos' => $this->urlDocumentos,
+            'filters' => $criteria,
         ]);                
     }
 
-    /**
-     * @Route("/{_locale}/regexpedientes", name="regexpedientes_index")
-     */
+    #[Route(path: '/{_locale}/regexpedientes', name: 'regexpedientes_index')]
     public function index(Request $request): Response
     {
         $this->loadQueryParameters($request);
